маленькие фиксы тестов
