+++ tests/test_bot.py
@@ -1,9 +1,9 @@
 from unittest.mock import patch
 
 
-def test_get_new_message(bot_fixture, ping_data):
+def test_get_new_message(bot_fixture, ping_data_fixture):
     with patch('requests.Session.request') as mock_get:
-        mock_get.return_value.json.return_value = ping_data
+        mock_get.return_value.json.return_value = ping_data_fixture
 
         @bot_fixture.handler.message_new()
         def response(message):
